requirements.txt
numpy>=1.24
opencv-python>=4.8
Pillow>=10.0
pytest>=7.0
pytest-mock>=3.0
ruff==0.0.291  # werkt met Python 3.11
tensorflow==2.14